make own compilation unit for panasonic decoders
